v0.2 UPDATE
Code developed completely, including GF calculation and Observables.
It still lack extension of the charge and pair susceptibilities.

STATUS:
comparison for the single band serial Bethe lattice is excellent at
last loop. (we solve the last loop of a reference for U=2).
Results are identical for the observables.
Matsubara GF are identical in 1d-12 (the threshold of the Arpack).
Real GF are identical which a larger error, but a visual check shows
that they look pretty identical on the real-axis.

Ready to extend to the full self-consistent calculation and
then to the multi band case.