v0.4 UPDATE - MANY BUGS FIX
This version of the code support a correct and working MPI
implementation, either with storage or not of the Hamiltonians
matrices.

The tensor structure of the Nup-Ndw basis required quite a
revision of the MPI structure.
Code has been tested successfully in teh BHZ-2D case.
Results are identical to serial case and within the same error with
respect to the former code.

The MPI structure is not optimal and some improvement might be required.
A rapid test of the scaling however shows good behavior with ed_sparse_H=False.
(sub-linear with 4cpu).

More tests to go.
Ready for beta-testing phase.