Development of the code:
Implemented the spreading of the eigen-vector among processors.
Rather than having each thread storing a copy of the eigenvector(s),
we let the threads save their chunk of the vector(s). These are
the ones actually used in the Matrix-Vector product, so most of the
time there is no need to recollect them.
In the GF calculation recollection is required because there is a non-local
operation on the vector. This is done by extending ED_EIGENSPACE to be
MPI-aware and using an MPI version of the function es_get_cvector, which
has Mpi communicator as first input.
This function return the recollected vector directly.

Notably, this function works with ED_TWIN.