UPDATE:
fixed!
This implementation has all the channels (local, non-local, DW, UP) sparse
matrices reduced in the MPI version.
This however requires to take the following steps, which are the major changes
with respect to the previous implementation:
- allocate a vout vector of dimension of the sector (Dim) to store
the result of the H_{up,dw}xV which is out-of-bound with respect to the
size of the vector chunks used in the parallel Lanczos
- all reduce the Vout into Vin (to save memory) before spread it among
threads.

Now ready to import the updated version of the Sparse Matrix, which:
+ divides the sparse matrix into a local - non-local part, this can
be used in principle to separate the matrix vector product in a part which
is local in memory and one that is not.
+ implements MPI manipulation directly, so one does not need to make them
externally.