v0.5 UPDATE:
This version implements a novel and more symmetric MPI scheme.
All the different parts of the sector sparse matrix are divided
into chunks. The diagonal/non-diagonal part is treated as before.
The UP/DW parts, which are defined thru a tensor product, are
treated differently as these do not have in general the same
start/end states in the full sector, i.e.
first_state_up/dw != first_state
last_state_up/dw  != last_state

This required to modify the matrix vector product, by reducing on a
tmp output vector of dimension Dim of the sector (large).
The gain is that the matrices are divided among the processors and
are in principle very small: 1/Ncpu*binom(Ns Nup/dw).

The SPARSE MATRIX module has been deeply updated with the improvements
tested in the OFFICINA/ARPACK_DETACH. So far the division between local/non-local
parts of the matrix are hidden, which makes the differentiation of many procedures
for the sparse matrices between MPI and serial quite useless.
See for example the insert_element one.

The next step would be to store the matrices using ELL format, rather than LL.
This guarantees access O(1) to the matrix elements rather than O(logN).
Yet, for very large systems not storing H should be the choice if the
matrix-vector product performs well enough (which I do not believe).