v0.6.0 UPDATE:
Code updated with support for *static* (as opposed to linked-list)
format for the sparse matrix. This format, which is the most similar
to the linked list one, provides access O(1) to matrix elements rather
than O(logN) as for the Linked-List.

The sparse matrix is stored as an array of array (as opposed to an
array of linked list) in a suitable data structure.
The first array has teh dimension of the number of rows needed.
The second array contains as many columns/values as required.
The number of them is estimated by a dry-run of the matrix construction.

Code tests serial/MPI for Hubbard1Band and BHZ. Results agrees within 1d-12
the Lanczos threshold.

>> THIS CONCLUDES THE FIRST ROUND OF UPDATES OF THE CODE <<

More tests to come.
Scaling analysis and performance.

Ready to deliver.