NEW VERSION
This version implements the separate conservation of the
quantum numbers Nup(iorb)-Ndw(iorb) per orbital.
The input parameter to control di execution mode is:
ed_total_ud = T (default) for total Nup-Ndw conservation
              F for per orbital Nup(a)-Ndw(a) conservation.

The Nup(a)-Ndw(a) are conserved separately in absence of any term
in the Hamiltonian which mixes different orbital sectors, i.e.
Spin-Exchange, Pair-Hopping, inter-orbital hopping terms (local
hybridization). Then the only non-local terms come from
the hybridization of the impurities with the corresponding bath levels.

The code writing required quite a lot of changes to be introduces.
In particular:
- all the maps getSector,getDimUp,getDimDw,getNup,getNdw are now
replaces by suitable functions which algorithmically evaluate the
corresponding variables. This is of course an improvement of the code
structure.
-  many procedures are now separated in two version, one for any
value of the logical parameter ed_total_ud.
- the Hamiltonian is built using a separate procedure, stored
in ED_HAMILTONIAN_MATVEC/stored/H*_orbs.f90
>>> ACTHUNG: the Nup(a)-Ndw(a) execution mode is now limited to:
   + non MPI case
   + non Twin states (although implemented: to be tested, this is
   going to be a huge speedup given the possible huge number of
   sectors in this mode) (no ed_twin=T)
   + no Direct HxV mode (no ed_sparse_H=F)
   + no ELL format for the H storage (no ed_sparse_format=ELL)
   + obviously no Hybrid bath mode: (bath_mode=hybrid)

Further tests to come.