V0.9.0 UPDATE:
The code has been updated with state-of-the-art MPI parallelization of the
matrix vector product.
The parallel HxV productds exploits the structure given by Nup-Ndw to greatly simplify the
communication among processes, latency and caching effects.

The new procedures use a generic and robust MPI_All2AllV routine to transpose the input vector
Vin regarded as a matrix in the indices {i_up, i_dw}.

The Hamiltonian and the product HxV are divided in three parts:
- H_local (diagonal terms): the corresponding product is parallelized as usual and only use local vector, i.e.
vector chunks belonging to the given thread. No communication required.
- H_up (Up part): this contributes to the non-local part of H as 11_dw (x) H_up. Thus the corresponding product
is performed for a FIXED i_dw configuration. This corresponds to multiply H_up with a chunk of local vector (i_up,i_dw_fixed)
of dimension DimUp*(DimDw/MpiSize). This vector resides on each thread memory. So the product is local: no
communicatin involved!
- H_dw (Dw part): this contributes to the non-lcoal part of H as H_dw (x) 11_up. Thus the corresponding product
is performed for a FIXED i_up configuration. This corresponds to multiply H_dw with a chunk of non-local vector (i_up_fixed,i_dw)
of dimension (DimUp/MpiSize)*DimDw. We can transform this product into a local one (contiguous in memory) by transposing the
chunk of vector (i_up,i_dw) --> (i_dw,i_up). This is performed using MPI communication with All2AllV.

The scaling of such HxV product was tested to be sub-linear in the speedup, as originally reported in the works of A.Dolfen and E.Koch.
Our tests was performed in a single eigenvalue calculation of a sparse Hamiltonian for a SIAM model, both with Arpack and Lanczos.
Testing in the ED code will be in order.

Additionally:
The use of memory has been largely improved by splitting the vectors used in the Lanczos calculations everywhere (each thread keep
only its own chunk and chunks of eigenvectors are stored separately). The master eventually gather/scatter the eigenvector
when application of any operator O|eigenvector> is required (GF calculation).

TODO:
The sparse matrix structure, currently using linked lists, should be updated into a dynamical vector structure (one
growing vector for any row). This will give O(1) access to the matrix elements rather than O(logN). And it is very
simple and cheap.