v0.11.1 BUG FIX
Fixed a bug in the new data structure in ED_SPARSE_MATRIX.
The stored rows are not sorted in the columns, this caused
a wrong exit of the binary_search algorithm when used
to locate an existent value for a given columns (iadd=.true.).

Solved by adding a quick sort algorithm inside binary_search.
The correct position is return using the sorting order.