v0.11.3 BUG FIX
The bug (noted by Hongchul Choi [SISSA]) that plagues the MPI executions of the code in sectors with just too few columns (DimDw < MpiSize) has been fixed.
The solution required us to split the communicator so to have a sub-group
of non-idle processors of the same size of DimDw (MpiSize_Local == DimDw).
This is achieved by:
1. setting a general MPI communicator and variables across the code (in ED_VARS_GLOBAL)
2. creating the sub-group in MPI at each sector building call (build_Hv_Sector)
3. freeing the sub-group in delete_Hv_sector.

In this way the code remains essentially unchanged, because all the work is done
in build/delete_Hv_sector.

Tests for Ns=10 using BHZ show that this fix does not change the solution (within
Lanc error threshold) notwithstanding the split of MPI in sectors of the form
(Nup=any,Ndw=0).

More tests to come.
This is yet a beta-version.