v0.11.4 BUG and UPDATE
Bugs fixed:
during tests runs we found a bug related to the gather/scatter of the eigenvectors.
If the GS sector was solved with full-diagonalization (eigh), because it is smaller
then the Lanczos threshold, the eigenvector were erroneously saved as full arrays in
the MPI mode, rather than as chunks (as in Lanczos). This caused a breakdown in the
call to es_return_cvec at the GF calculation, as stop condition (Ndim != Dim) was
met.
Fixed by scattering the basis to the threads in the MPI mode.

Update:
Introduced a flag "ed_sectors" and "ed_sectors_shift" to drastically and dynamically
reduce the sectors scan.
IF ed_sectors=T the code reads the file "sectors_list.restart" containing Indices for the
actual GS of the previous loop. Based on this information it creates a list of sectors
to be visited add/removing "ed_sectors_shift" particles in all channel.
Tested with a "quench" in the occupation search (starting from a wrong chemical potential
and letting the code determing the correct one for a much smaller occupation)