Intermediate commit
Added simple Lanczos solver for the groundstate at T=0.
Follows from critical bug fix in SCIFOR and DMFT_ED generic code.

Results are numerically identical between lanczos and arpack solver here
and with respect to the DMFT_ED code.

This commit will be merged in master, but SE and PH keep NOT WORKING.