v0.12.0 UPDATE & CRITICAL BUG FIX
This version includes the full rotational invariant Kanamori interaction,
with Spin-Exchange and Pair-Hopping terms.

The code has been tested successfully with respect to the generic (older)
code in a 2-bands model, with full interaction. More tests are in order.

The inclusion of Kanamori interaction is made possible by using a sparse
non-local hamiltonian, which however requires to modify the nice and fast
MPI product. This can potentially destroy the nice scaling in the MPI mode,
but, hey, there are no free lunches...

BUG FIX: CRITICAL
The dumping of the small sparse matrices Hup, Hdw into dense matrices
thru a tensor product (see ED_HAMILTONIAN_SPARSE_HxV) in order to
perform a direct Lapack diagonalization was wrong.
The order of the matrices in the tensor product had to be exchanged
 Hmat = Hmat + kronecker_product(Htmp_dw,eye(DimUp))
 Hmat = Hmat + kronecker_product(eye(DimDw),Htmp_up)

 H_DW _circ_ 1_UP +  1_DW _circ_ H_UP

rather than:

 H_UP _circ_ 1_DW +  1_UP _circ_ H_DW

which gave incorrect results for the small sectors digonalized exactly.

Other improvements:
- include a orbital selective fit + a routine to impose orbital equality
of the bath (fit one get all).

- updated the default and comments in ED_INPUT_VARS.f90