v0.12.0 UPDATE & BUG FIX
This version includes the full rotational invariant Kanamori interaction,
with Spin-Exchange and Pair-Hopping terms.

The code has been tested successfully with respect to the generic (older)
code in a 2-bands model, with full interaction. More tests are in order.

The inclusion of Kanamori interaction is made possible by using a sparse
non-local hamiltonian, which however requires to modify the nice and fast
MPI product. This can potentially destroy the nice scaling in the MPI mode,
but, hey, there are no free lunches...

Other improvements:
- include a orbital selective fit + a routine to impose orbital equality
of the bath (fit one get all).