v0.12.1 UPDATE
Added several procedures in ED_IO and ED_BATH_FUNCTIONS to retrieve the
non-interacting Anderson functions (G0and, invG0and, Delta).
There are two sets of functions:
1. ed_get_g0imp_{matsubara,realaxis}: these works just like the functions
that return the self-energies of impurity Green's funtions. in ED_IO

2. ed_{delta,g0imp,invg0}_{matsubara,realaxis}: these accept an array of
frequencies and the user bath (double precision array as passed or to be
passed to the ed_solve). They return the Delta(z), calG0(z) or
inv{calG0}(z) on either matsubara or realaxis. in ED_BATH_FUNCTIONS