v0.12.9 Update and new features
- Added a novel input variable: *cg_minimize_hh* to select the
hh parameter in the CG minimize subroutine.

- Added a novel input variable: *cg_minimize_ver* to select which
version of the minimize subroutine to use.
There are two versions:
 + Krauth = OLD/.FALSE.
 + Sascha = NEW/.TRUE.

The new routine seems to be improved with respect to the old one.
The default value of the new variable is .FALSE. to
ensure back-compatibility.
We tested that the parameter cg_minimize_hh governs somehow
the quality of the fit, which is however identical (for a
fixed hh) among the two subroutines.
Of course more tests are in order to better understand the
possible different behavior.

The tests are performed in a single run from a given guess
in a three bands Bethe lattice model calculation.

So far we have:
Sascha==Krauth==Num.Rec. for hh=1d-4
Sascha==Krauth/=Num.Rec. for hh=1d-5