v0.12.11 CRITICAL (BUT MOSTLY UNNOTICED) BUG FIX
A test in a parent code (VCA) with LC spotted the
presence of a critical bug, fortunately almost unnoticed.
The bug concerns the creation of the map get{C,CDG}sectors
in the setup_global procedure.

A wrong loop arrangement and if(true/false)cycle conditions
caused these array to be left incomplete for extremal sectors,
i.e. for sectors which have quantum numbers of the form
nup {0,Ns} or {0,Ns} ndw

In this case the if(Jup/dw<0)cycle, if(Jup/dw>Ns)cycle was met,
causing the restart of the loop and leaving the arrays below the
given condition blank.
In a structure like the following, if first condition is met (jup>0)
then not only A is not performed, but automatically also B,C,D.

do
 if(jup<0)cycle
 do A

 if(jup>Ns)cycle
 do B

 if(jdw<0)cycle
 do C

 if(jdw>Ns)cycle
 do D

Solved by repeating the loop over all the get{C,CDG}sector(:,1/2,isector) cases.