0.13.0 Bug fix in MPI + Updates
The code has been updated following the lastest development in Scifortran
This realease breaks compatility. Update of Scifor is mandatory.
Note that the updated is just a matter of required/optional variables
in the arpack/lanczos interface.

A possible error occurred in MPI execution if the following conditions
are met:
- A sector solved by EIGH (lapack) is found
- Ncpu is larger than the MPI-split of that sector dimension
This causes the sector to be solved by a reduced number of MPI_UNIVERSE.

However a misplacement of  *delete_Hv_sector()* just after diagonalization
caused the scatter of the resulting eigen-vectors with a different,
larger, MPI_UNIVERSE. This caused a freeze out of the code execution,
due to a mismatch in the communication.

Apparently solved moving *delete_Hv_sector()* just below the save/scatter
eig-vectors part. See ED_DIAG.f90

To be tested on the HPC cluster.