0.13.1 Update + Patch no-fix Mpi_all2all hang bug
In the previous commit ("5397558bb9cdab148b4276ed121c2f21fa18fa7e")
I claim to have solved a nasty but misterious bug. This is unfortunately
not true. The bug is still there.
The bug is a hanging of the executon for the sector
  28, nup:  2, ndw:  1, dims=    66    12                 792
> 29, nup:  2, ndw:  2, dims=    66    66                4356 <

in a Bethe, Norb=2, Nbath=5 (12 levels) executed with Ncpu>12 and lanc_dim_threshold < 792

I trace back the bug to the hanging of the first call to mpi_all2all in the
mpi_transpose procedure. This call is required to populate the recv_counts
buffer, so it is very simple. Yet the code enters there and never exit.
I suspect a bug in openmpi (we have not noticed such a bug in Mpich2 so far).

The bug is certainly related to some mismatch in the internal MPI communicators.
In fact it is simply cured by changing the lanc_dim_threshold so to have the
sector 28 be solved exactly (lanc_dim_threshold > 792).
I guess the problem is related to the MPI reduction executed in sector 28, which is
by the way correctly solved.
The communicators in the sector 29 are anyway well defined. I did test some cleaning
of the creation/destruction of the MPI comms with no success.

Getto la spugna cazzo... I am too old for this shit!