Intermediate commit: Debug TOTAL_UD=F
There is an inconsistency in the diagonalization of the H matrix in some
small sector that we noticed for Norb=2 Nbath=2.
We analyzed the Hamiltonian structure H=H_d + sum_a kp(1....h_a...1)
and everythin fitted our hand made calculations in the sector
0 0 1 2 (dims=1 1 3 3 = 9)

Notably, similar sector (0 0 1 1) with same dimensions give eigenvalues taht
are in agreement with the Nup,Ndw algorithm. WHy is this sector failing to get
correct Eigenvalues?