Intermediate commit.
Solved a major bug in ED_SETUP, getCDGsector, getCsector when operating
with Nups,Ndws orbitally resolved. There was a typo: using Ns rather than
Ns_orb to check the limits of the two maps. This led to operate in wrong
sector when applying c,cdw operators eventually getting zero vectors and
crashing the lanczos evaluation of the GF.

Still problems with handling dense matrices in the ed_total_UD=F.

What about removing Lapack in Lanczos? in the end we do not need it
anymore.