Debug complete
Debug of the Chi(z) functions for the spin-spin, dens-dens cases are
completely debugged. In collaboration with A.Scazzola.

Note that at T=0 the Chi(tau) needs not to go back to its value at tau=0
for tau = beta. This is required by periodicity at finite Temperature,
but for T=0, beta--->infty.

We tested that at finite T with Lanczos and for the limit of
NGFITER >> 100, the correct behavior at finite temperature is
recovered, although many many alfas,betas are required. This is
reasonable since the inversion by tridiagonalization of H is an
approximation. No matter what in the 0:beta/2 regime the Chi are
correct.

Ready to merge with +master branch