0.18 BUG FIX
Fixed a bug in the H_non_local part.
The full MPI decomposition we use for matrices H_local and H_non_local is based
on the MPI decomposition of the DW part only. This full MPI is working for the
old *transposed* construction of the Hamiltonian but it is not for the new direct
one. Now H_non_local is symmetric so there is no difference among the two matrices
(transposed and direct). We exploit this fact to use the full MPI decomposition
in constructing the transposed Hamiltonian rather than the direct.

Perhaps in the future we will come up with a better full MPI decomposition.
Note however that this is a problem only for the non-local part (Jx, Jp).