0.18.1 BUG FIX
Fixed (or so) a couple of bugs discovered by A.Scazzola.
- vvinit in ED_GF_XXX was allocated only by the master, which is in charge of
applying operators to the GS and only then scatter it to the other nodes.
Yet calling the MPI_scatter procedure with no allocated vector from the other
nodes generates an error in some debugging mode (e.g for intel). To follow the
correct pattern we let the other nodes allocate an empty vvinit (vvinit(0)).

- A more serious but unavoidable bug is present in MPI_TRANPOSE (ED_HAMILTONIAN_COMMON).
Here the all_to_all_v communication is performed column by column by the MPI universe using
the largest (Ntot) number of columns. So, some node participate in the communication with
an empty vector (A(:,j) with j>Qcol_node). Normally this produce an empty communication (
no data to pass). With some debug options (in Intel -check) this produce an error message
during execution. We can not at this stage fix this bug. So we reduced the level of debug
by using -no-check in Intel FC.
Note that with GNU no problems arise (apparently).