0.19.0 NEW VERSION
This updated version includes a novel replica bath form, largely inspired
by that used in the CDMFT work where it has been developed and tested.
We introduced this bath because:
- It generalized the previous version by introducing the possibility of
specifying the local Hamiltonian matrix basis, i.e. a basis (loosely
speaking) which decomposes the local Hamiltonian.
- It allows to "easily" get the gradient of the Delta function
V*(z-H)**-1*V

This updated version also includes the analytic calculation of the
gradient of the Delta and Weiss/G0and functions used in the fit.
Using the analytic gradient rather than the numerical one possibly
ensures to have a better stability of the fit/solution near unstable
regions.
The gradient has been tested successfully for normal, hybrid calculations.
For the replica bath it seems to take such a long time that makes the
fit impossible, i.e. never exiting. We will work to fix this issue.