0.19.2 UPDATE + BUG FIX
The code has been updated to comply with the latest development in
SciFor. The conjuate gradient has now only one tolerance parameter
(tol --> fixed by cg_tol) which is used to tune the accuracy of the
calculation.

Some adjustments have been done in the FIT part.
A possible bug has been removed.
The gradient of the Delta/Weiss fittin functions used in the
chi**2 was correct only for the cg_pow=2 case. This is now
generalized to any cg_pow case.

>> Note that the actual form of the analytic gradient
for the replica bath it is only working for diagonal orbitals, i.e.
no local hybridization. A fix has been found but not yet coded.
<<