0.19.4 POSSIBLE BUG FIXED
A.SCAZZOLA found a possible bug in the I/O of the bath parameters.

The bath parameters were written using a definite F21.12 format.
In case the parameters became smaller than 1D-12 they appeared as
zero (identically) in the file, which when used to restart the
simulation led to an improper vanishing of the analytic gradient.

Solved by changing the format for the output of the bath parameters
as ES21.12.
Tested for the 2bands Hubbard Bethe.

The hamiltonian.xxx files are now slightly different from before (use
exp notation) but reading part is free format, so no problem should
arise with back compatibility.