0.19.5 BUG FIXED
A misbehavior introduced with one of the latest update, i.e.
the T=0 Full ED, has been fixed.
The T=0 Full ED method based on a change of the energy cutoff proved
to be unstable in some situation. A more efficient and memory saving
procedure is to use a mixed method Lapack + Lanczos.
If T=0 full ED is required the lanc_dim_threshold parameter, fixing
the largest dimension solved with Lapack rather than Lanczos, is
increased up to the dimension of the largest sector. IN this way all
sectors are exactly diagonalized. Yet, only the groundstate(s) is(are)
retained in the state_list data structure. The rest of the code
proceeds as usual: GF is obtained via Krylov method and observables
are calculated as usual.
Mathematically this is equivalent to the other, but it has the a
advantage of using far less memory and be faster.