Changed all the necessary files to run the parallel version of the code with phonons (mainly added some loops, the distinction between phonon and electron indices, some other small things... everything is similar to the serial part but with different dimensions)
ED_GF_NORMAL, _CHIDENS, _CHISPIN: there was a vector allocated to 0
elements, now it is allocated to 1 elements (= 0) to avoid problems
inside the subroutine scatter_vector_MPI (inside ED_SETUP)

ED_HAMILTONIAN/direct_mpi/HxV_non_local.f90: changed the way the product
is performed (exchanged two indices, i.e. we build the transpose matrix
instead of the correct one, but it is not a problem since the matrix is
symmetric), in order to avoid an index going out of boundaries. This was
a problem NOT related to phonons.