intermediate update.
+ Major reorder  of the GF and Chi modules.
+ Added pair susceptibility (shame on me)
+ Coding excitonic susceptibility part

++ Major change: introduced *sector* type in
VARS_GLOBAL to keep track of all the sector info.
Use that to abstract the application of operators to
Fock state: Op|i> = sgn|j> in SETUP.
These will be used to largely simplify the code
at GF and Chi evaluation, which in some parts is now a mess whereas
the theory is pretty simple indeed.